Fix PID speed unit conversion bug (int vs float)

+++ arduino/Wheel.h
@@ -10,7 +10,7 @@ class Wheel {
     /**
      * Frequency (in Hertz) between consecutive calls to #updateOdometry().
      */
-    const int mUpdateFrequency;
+    const float mUpdateFrequency;
     
     /**
      * PID coefficients.
